requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1
django-environ==0.4.5
django-import-export==2.3.0
django-ip-geolocation==1.6.0
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
fcm-django==0.3.4
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
Pillow==8.1.0
psycopg2-binary==2.8.5
pyfcm==1.4.7
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
razorpay==1.2.0
requests==2.24.0
simplejson==3.17.2
sqlparse==0.3.1
tablib==2.0.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.10
user-agents==2.2.0
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
yapf==0.30.0
